requirements.txt
Flask
whisper
openai-whisper
pydub
python-docx
pdfplumber
transformers
torch
spacy
